pyproject.toml
[tool.poetry]
name = "oracles"
version = "0.2.0"
description = ""
authors = ["Ben Steinberg <ben@potswift.org>"]
readme = "README.md"
license = "GPL-3.0-or-later"
packages = [{include = "oracles"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.2"
click = "^8.1.3"

[tool.poetry.group.dev.dependencies]
hypothesis = "^6.60.0"
pytest-playwright = "^0.3.0"
pytest-env = "^0.8.1"
pytest-flask = "^1.2.0"
flake8 = "^6.0.0"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
oracles = 'oracles.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=oracles --live-server-port=5000"
env = [
    "ORACLES=data_sample",
    "FLASK_DEBUG=1"
]
pythonpath = [
    "oracles"
]

[tool.coverage.run]
omit = ['oracles/data.py']